requirements.txt
django-cms
djangocms_plus